+++ CHANGELOG.md
@@ -5,6 +5,7 @@ V1.5.0
 -------
 * Added command: recentTasks (csvd) - Output CSV of items tracked within current day
 * Added command: recentTasks (csvt) - Output CSV of 30 most recently tracked task numbers
+* Improved (simplified) output of "w" command when report is empty 
 
 V1.4.1
 -------